Python/E11/requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
Pillow==8.3.2
